### 2021-02-16 4faa507
Update dois and references
