+++ Docs/paper/paper.bib
@@ -1,35 +1,48 @@
 @ARTICLE{onorbe2019,
-       author = {{O{\~n}orbe}, Jose and {Davies}, F.~B. and {Luki{\'c}} and {}, Z. and {Hennawi}, J.~F. and {Sorini}, D.},
-        title = "{Inhomogeneous reionization models in cosmological hydrodynamical simulations}",
-      journal = {\mnras},
-     keywords = {methods: numerical, intergalactic medium, dark ages, reionization, first stars, early Universe, large-scale structure of Universe, Astrophysics - Cosmology and Nongalactic Astrophysics},
-         year = 2019,
-        month = jul,
-       volume = {486},
-       number = {3},
-        pages = {4075-4097},
-          doi = {10.1093/mnras/stz984},
-archivePrefix = {arXiv},
-       eprint = {1810.11683}
+    author = {{O{\~n}orbe}, Jose and {Davies}, F.~B. and {Luki{\'c}} and {}, Z. and {Hennawi}, J.~F. and {Sorini}, D.},
+    title = "{Inhomogeneous reionization models in cosmological hydrodynamical simulations}",
+    journal = {Monthly Notices of the Royal Astronomical Society},
+    volume = {486},
+    number = {3},
+    pages = {4075-4097},
+    year = {2019},
+    month = {05},
+    issn = {0035-8711},
+    doi = {10.1093/mnras/stz984},
+    url = {https://doi.org/10.1093/mnras/stz984},
+    eprint = {https://academic.oup.com/mnras/article-pdf/486/3/4075/28559361/stz984.pdf}
 }
 
 
 @article{nyx-paper1,
-  title={Nyx: A Massively Parallel AMR Code for Computational Cosmology},
-  author={A.S. Almgren adnd J.B. Bell and M.J. Lijewski and Z. Lukic and E. Van Andel},
-  journal={Astrophysical Journal},
-  volume={765},
-  number={39},
-  year={2013},
-  publisher={IOP Publishing}
+	doi = {10.1088/0004-637x/765/1/39},
+	url = {https://doi.org/10.1088/0004-637x/765/1/39},
+	year = 2013,
+	month = {feb},
+	publisher = {American Astronomical Society},
+	volume = {765},
+	number = {1},
+	pages = {39},
+	author = {Ann S. Almgren and John B. Bell and Mike J. Lijewski and Zarija Luki{\'{c}} and Ethan Van Andel},
+	title = {Nyx: A {MASSIVELY} {PARALLEL} {AMR} {CODE} {FOR} {COMPUTATIONAL} {COSMOLOGY}},
+	journal = {The Astrophysical Journal},
+	abstract = {We present a new N-body and gas dynamics code, called Nyx, for large-scale cosmological simulations. Nyx follows the temporal evolution of a system of discrete dark matter particles gravitationally coupled to an inviscid ideal fluid in an expanding universe. The gas is advanced in an Eulerian framework with block-structured adaptive mesh refinement; a particle-mesh scheme using the same grid hierarchy is used to solve for self-gravity and advance the particles. Computational results demonstrating the validation of Nyx on standard cosmological test problems, and the scaling behavior of Nyx to 50,000 cores, are presented.}
 }
 
 @article{maestroex,
-  title={MAESTROeX: A Massively Parallel Low Mach Number Astrophysical Solver},
-  author={D. Fan and A. Nonaka and A. S. Almgren and A. Harpole and M. Zingale},
-  journal={accepted for publication in The Astrophysical Journal},
-  year={2019},
-  url = {https://arxiv.org/abs/1908.03634}
+	doi = {10.3847/1538-4357/ab4f75},
[…]
