### 2021-02-26 #3068 @whedon
[REVIEW]: Nyx: A Massively Parallel AMR Code for Computational Cosmology
**Submitting author:** @jmsexton03 (<a href="http://orcid.org/0000-0003-2551-1678">Jean M. Sexton</a>)
**Repository:** <a href="https://github.com/AMReX-Astro/Nyx" target ="_blank">https://github.com/AMReX-Astro/Nyx</a>
**Version:** 21.07
