### 2021-03-01 ed79cac
Add strang restart from sdc workaround
