+++ Source/Hydro/Godunov.cpp
@@ -69,10 +69,8 @@ pc_umeth_3D(
   int cdir = 0;
   const amrex::Box& xmbx = growHi(bxg2, cdir, 1);
   const amrex::Box& xflxbx = surroundingNodes(grow(bxg2, cdir, -1), cdir);
-  amrex::FArrayBox qxm(xmbx, QVAR);
-  amrex::FArrayBox qxp(bxg2, QVAR);
-  amrex::Elixir qxmeli = qxm.elixir();
-  amrex::Elixir qxpeli = qxp.elixir();
+  amrex::FArrayBox qxm(xmbx, QVAR, amrex::The_Async_Arena());
+  amrex::FArrayBox qxp(bxg2, QVAR, amrex::The_Async_Arena());
   auto const& qxmarr = qxm.array();
   auto const& qxparr = qxp.array();
 
@@ -80,10 +78,8 @@ pc_umeth_3D(
   cdir = 1;
   const amrex::Box& yflxbx = surroundingNodes(grow(bxg2, cdir, -1), cdir);
   const amrex::Box& ymbx = growHi(bxg2, cdir, 1);
-  amrex::FArrayBox qym(ymbx, QVAR);
-  amrex::FArrayBox qyp(bxg2, QVAR);
-  amrex::Elixir qymeli = qym.elixir();
-  amrex::Elixir qypeli = qyp.elixir();
+  amrex::FArrayBox qym(ymbx, QVAR, amrex::The_Async_Arena());
+  amrex::FArrayBox qyp(bxg2, QVAR, amrex::The_Async_Arena());
   auto const& qymarr = qym.array();
   auto const& qyparr = qyp.array();
 
@@ -91,10 +87,8 @@ pc_umeth_3D(
   cdir = 2;
   const amrex::Box& zmbx = growHi(bxg2, cdir, 1);
   const amrex::Box& zflxbx = surroundingNodes(grow(bxg2, cdir, -1), cdir);
-  amrex::FArrayBox qzm(zmbx, QVAR);
-  amrex::FArrayBox qzp(bxg2, QVAR);
-  amrex::Elixir qzmeli = qzm.elixir();
-  amrex::Elixir qzpeli = qzp.elixir();
+  amrex::FArrayBox qzm(zmbx, QVAR, amrex::The_Async_Arena());
+  amrex::FArrayBox qzp(bxg2, QVAR, amrex::The_Async_Arena());
   auto const& qzmarr = qzm.array();
   auto const& qzparr = qzp.array();
 
@@ -186,11 +180,9 @@ pc_umeth_3D(
   // These are the first flux estimates as per the corner-transport-upwind
   // method X initial fluxes
   cdir = 0;
-  amrex::FArrayBox fx(xflxbx, flx1.nComp());
-  amrex::Elixir fxeli = fx.elixir();
+  amrex::FArrayBox fx(xflxbx, flx1.nComp(), amrex::The_Async_Arena());
   auto const& fxarr = fx.array();
-  amrex::FArrayBox qgdx(xflxbx, NGDNV);
-  amrex::Elixir qgdxeli = qgdx.elixir();
+  amrex::FArrayBox qgdx(xflxbx, NGDNV, amrex::The_Async_Arena());
   auto const& gdtempx = qgdx.array();
   amrex::ParallelFor(
     xflxbx, [=] AMREX_GPU_DEVICE(int i, int j, int k) noexcept {
@@ -205,11 +197,9 @@ pc_umeth_3D(
 
   // Y initial fluxes
   cdir = 1;
-  amrex::FArrayBox fy(yflxbx, flx2.nComp());
[…]
