### 2021-03-04 AMReX-Astro/Nyx#76 @c-white
JOSS Review: Statement of Need
I am one of the reviewers for the [JOSS submission](https://github.com/openjournals/joss-reviews/issues/3068). Everything looks great so far, though I have a tiny suggestion regarding the statement of need. It could be a bit more specific about:

1. "What problems the software is designed to solve" (though I understand that's mentioned elsewhere in the paper).
2. More importantly, "who the target audience is." For example, the code seems aimed at researchers with access to at least some HPC resources (as opposed to, say, students taking their first cosmology class).
