### 2021-03-04 #3068 @c-white
@dfm I have finished my review, and everything seems fine to me. Some notes:

1. **Performance:** Everything installed (remarkably smoothly!) and I was able to run some basic parallel simulations on a modest cluster. This doesn't cover massive simulations, or GPU clusters for that matter, but I have no reason to believe things won't work in that regime.
2. **Functionality documentation:** I don't know if you're looking for documentation of every function in the code base, which there is not, but I don't think that's quite necessary for a code this size. The documentation does thoroughly describe user inputs, as well as the underlying equations and algorithms employed.
3. Maybe I don't understand whedon, but the pdf it just generated doesn't have the changes the authors recently pushed to the paper.
