### 2021-03-05 #3068 @dfm
@c-white: Thanks for this! Your interpretation of the performance and API docs requirements sounds good.

For the paper, I am seeing [the most recent changes](https://github.com/AMReX-Astro/Nyx/commit/048a26bdae232868105444685180816617b8ba3e) at the link above: https://github.com/openjournals/joss-reviews/issues/3068#issuecomment-790986670 Are there other commits that are missing? Thanks again!
