+++ Source/Hydro/Godunov.cpp
@@ -69,8 +69,10 @@ pc_umeth_3D(
   int cdir = 0;
   const amrex::Box& xmbx = growHi(bxg2, cdir, 1);
   const amrex::Box& xflxbx = surroundingNodes(grow(bxg2, cdir, -1), cdir);
-  amrex::FArrayBox qxm(xmbx, QVAR, amrex::The_Async_Arena());
-  amrex::FArrayBox qxp(bxg2, QVAR, amrex::The_Async_Arena());
+  amrex::FArrayBox qxm(xmbx, QVAR);
+  amrex::FArrayBox qxp(bxg2, QVAR);
+  amrex::Elixir qxmeli = qxm.elixir();
+  amrex::Elixir qxpeli = qxp.elixir();
   auto const& qxmarr = qxm.array();
   auto const& qxparr = qxp.array();
 
@@ -78,8 +80,10 @@ pc_umeth_3D(
   cdir = 1;
   const amrex::Box& yflxbx = surroundingNodes(grow(bxg2, cdir, -1), cdir);
   const amrex::Box& ymbx = growHi(bxg2, cdir, 1);
-  amrex::FArrayBox qym(ymbx, QVAR, amrex::The_Async_Arena());
-  amrex::FArrayBox qyp(bxg2, QVAR, amrex::The_Async_Arena());
+  amrex::FArrayBox qym(ymbx, QVAR);
+  amrex::FArrayBox qyp(bxg2, QVAR);
+  amrex::Elixir qymeli = qym.elixir();
+  amrex::Elixir qypeli = qyp.elixir();
   auto const& qymarr = qym.array();
   auto const& qyparr = qyp.array();
 
@@ -87,8 +91,10 @@ pc_umeth_3D(
   cdir = 2;
   const amrex::Box& zmbx = growHi(bxg2, cdir, 1);
   const amrex::Box& zflxbx = surroundingNodes(grow(bxg2, cdir, -1), cdir);
-  amrex::FArrayBox qzm(zmbx, QVAR, amrex::The_Async_Arena());
-  amrex::FArrayBox qzp(bxg2, QVAR, amrex::The_Async_Arena());
+  amrex::FArrayBox qzm(zmbx, QVAR);
+  amrex::FArrayBox qzp(bxg2, QVAR);
+  amrex::Elixir qzmeli = qzm.elixir();
+  amrex::Elixir qzpeli = qzp.elixir();
   auto const& qzmarr = qzm.array();
   auto const& qzparr = qzp.array();
 
@@ -180,9 +186,11 @@ pc_umeth_3D(
   // These are the first flux estimates as per the corner-transport-upwind
   // method X initial fluxes
   cdir = 0;
-  amrex::FArrayBox fx(xflxbx, flx1.nComp(), amrex::The_Async_Arena());
+  amrex::FArrayBox fx(xflxbx, flx1.nComp());
+  amrex::Elixir fxeli = fx.elixir();
   auto const& fxarr = fx.array();
-  amrex::FArrayBox qgdx(xflxbx, NGDNV, amrex::The_Async_Arena());
+  amrex::FArrayBox qgdx(xflxbx, NGDNV);
+  amrex::Elixir qgdxeli = qgdx.elixir();
   auto const& gdtempx = qgdx.array();
   amrex::ParallelFor(
     xflxbx, [=] AMREX_GPU_DEVICE(int i, int j, int k) noexcept {
@@ -197,9 +205,11 @@ pc_umeth_3D(
 
   // Y initial fluxes
   cdir = 1;
[…]
