+++ Source/HeatCool/integrate_state_with_source_3d.cpp
@@ -1,17 +1,7 @@
 #include <fstream>
 #include <iomanip>
 
-#include <AMReX_ParmParse.H>
-#include <AMReX_Geometry.H>
-#include <AMReX_MultiFab.H>
-#include <AMReX_Print.H>
-#include <AMReX_PlotFileUtil.H>
-
-#include <AMReX_BLFort.H>
-#include <Nyx.H>
-#include <f_rhs.H>
-#include <f_rhs_struct.H>
-
+#include <AMReX_SUNMemory.H>
 #include <cvode/cvode.h>               /* prototypes for CVODE fcts., consts. */
 #include <cvode/cvode_diag.h>          /* access to CVDiag interface */
 #include <sundials/sundials_types.h>   /* definition of type realtype */
@@ -29,6 +19,18 @@
 #ifdef AMREX_USE_DPCPP
 #include <nvector/nvector_sycl.h>
 #endif
+
+#include <AMReX_ParmParse.H>
+#include <AMReX_Geometry.H>
+#include <AMReX_MultiFab.H>
+#include <AMReX_Print.H>
+#include <AMReX_PlotFileUtil.H>
+
+#include <AMReX_BLFort.H>
+#include <Nyx.H>
+#include <f_rhs.H>
+#include <f_rhs_struct.H>
+
 //#define MAKE_MANAGED 1
 using namespace amrex;
 
@@ -150,67 +152,8 @@ int Nyx::integrate_state_struct_mfin
 
 #ifdef AMREX_USE_CUDA
       cudaStream_t currentStream = amrex::Gpu::Device::cudaStream();  
-      if(sundials_alloc_type%2==0)
-      {
-                if(sundials_alloc_type==0)
-                  u = N_VMakeWithManagedAllocator_Cuda(neq,sunalloc,sunfree);  /* Allocate u vector */
-                else
-                  u = N_VNewManaged_Cuda(neq);  /* Allocate u vector */
-
-                dptr=N_VGetDeviceArrayPointer_Cuda(u);
-
-                if(sundials_alloc_type==0)
-                  e_orig = N_VMakeWithManagedAllocator_Cuda(neq,sunalloc,sunfree);  /* Allocate u vector */
-                else
-                  e_orig = N_VNewManaged_Cuda(neq);  /* Allocate u vector */
-
-                eptr=N_VGetDeviceArrayPointer_Cuda(e_orig);
-                N_VSetCudaStream_Cuda(e_orig, &currentStream);
-                N_VSetCudaStream_Cuda(u, &currentStream);
-
-                if(sundials_alloc_type==0)
-                  abstol_vec = N_VMakeWithManagedAllocator_Cuda(neq,sunalloc,sunfree);  
-                else
-                  abstol_vec = N_VNewManaged_Cuda(neq);  /* Allocate u vector */
-
-                abstol_ptr = N_VGetDeviceArrayPointer_Cuda(abstol_vec);
-                N_VSetCudaStream_Cuda(abstol_vec,&currentStream);
-                amrex::Gpu::Device::streamSynchronize();
-      }
-      else
-      {
-                dptr=(double*) The_Arena()->alloc(neq*sizeof(double));
-                u = N_VMakeManaged_Cuda(neq,dptr);  /* Allocate u vector */
-                eptr= (double*) The_Arena()->alloc(neq*sizeof(double));
-                e_orig = N_VMakeManaged_Cuda(neq,eptr);  /* Allocate u vector */
-                N_VSetCudaStream_Cuda(e_orig, &currentStream);
-                N_VSetCudaStream_Cuda(u, &currentStream);
-
[…]
