### 2021-03-26 c1ee73e
Add DPCPP compile guard, change managed type for HIP
