### 2021-03-27 0e295fb
Prefer using non-managed HIP Sundials Vector, since we never access it on the host
