### 2021-03-29 bc0d91f
Add Real enforcement for PRECISION=FLOAT example
