### 2021-03-29 8f1ecbf
Add timer regions on reductions
