### 2021-03-29 ecdf298
Recharacterize default MLPoisson object reuse
