### 2021-04-09 8cc8aca
OMP hacking for sundials test
