+++ Source/HeatCool/integrate_state_with_source_3d.cpp
@@ -55,10 +55,10 @@ int Nyx::integrate_state_struct
   //#endif
 #ifdef _OPENMP
 #ifdef AMREX_USE_GPU
-    if (sundials_use_tiling)
+  //    if (sundials_use_tiling)
        const auto tiling = MFItInfo().SetDynamic(true);
-    else
-       const bool tiling = false;
+       //    else
+       //       const bool tiling = false;
 #pragma omp parallel
 #else
     const bool tiling = (TilingIfNotGPU() && sundials_use_tiling);
@@ -190,27 +190,33 @@ int Nyx::integrate_state_struct_mfin
                 abstol_vec = N_VMakeManaged_Cuda(neq,abstol_ptr);
                 N_VSetCudaStream_Cuda(abstol_vec,&currentStream);
                 amrex::Gpu::streamSynchronize();
+		
               }
-                        if(sdc_iter>=0||true)
-              {
-              T_vec = N_VMakeWithManagedAllocator_Cuda(neq,sunalloc,sunfree);
-              ne_vec = N_VMakeWithManagedAllocator_Cuda(neq,sunalloc,sunfree);
-              rho_vec = N_VMakeWithManagedAllocator_Cuda(neq,sunalloc,sunfree);
-              rho_init_vec = N_VMakeWithManagedAllocator_Cuda(neq,sunalloc,sunfree);
-              rho_src_vec = N_VMakeWithManagedAllocator_Cuda(neq,sunalloc,sunfree);
-              rhoe_src_vec = N_VMakeWithManagedAllocator_Cuda(neq,sunalloc,sunfree);
-              e_src_vec = N_VMakeWithManagedAllocator_Cuda(neq,sunalloc,sunfree);
-              IR_vec = N_VMakeWithManagedAllocator_Cuda(neq,sunalloc,sunfree);
-              }
-              amrex::Real* T_vode= N_VGetDeviceArrayPointer_Cuda(T_vec);
-              amrex::Real* ne_vode=N_VGetDeviceArrayPointer_Cuda(ne_vec);
-              amrex::Real* rho_vode=N_VGetDeviceArrayPointer_Cuda(rho_vec);
-              amrex::Real* rho_init_vode=N_VGetDeviceArrayPointer_Cuda(rho_init_vec);
-              amrex::Real* rho_src_vode=N_VGetDeviceArrayPointer_Cuda(rho_src_vec);
-              amrex::Real* rhoe_src_vode=N_VGetDeviceArrayPointer_Cuda(rhoe_src_vec);
-              amrex::Real* e_src_vode=N_VGetDeviceArrayPointer_Cuda(e_src_vec);
-              amrex::Real* IR_vode=N_VGetDeviceArrayPointer_Cuda(IR_vec);
-
+      amrex::Real* T_vode=(realtype*) The_Arena()->alloc(neq*sizeof(realtype));
+      T_vec = N_VMakeManaged_Cuda(neq, T_vode);
+      amrex::Real* ne_vode=(realtype*) The_Arena()->alloc(neq*sizeof(realtype));
+      ne_vec = N_VMakeManaged_Cuda(neq, ne_vode);
+      amrex::Real* rho_vode=(realtype*) The_Arena()->alloc(neq*sizeof(realtype));
+      rho_vec = N_VMakeManaged_Cuda(neq, rho_vode);
+      amrex::Real* rho_init_vode=(realtype*) The_Arena()->alloc(neq*sizeof(realtype));
+      rho_init_vec = N_VMakeManaged_Cuda(neq, rho_init_vode);
[…]
