+++ Source/HeatCool/integrate_state_with_source_3d.cpp
@@ -107,6 +107,7 @@ int Nyx::integrate_state_struct_mfin
    long int& old_max_steps, long int& new_max_steps,
    const int sdc_iter)
 {
+    BL_PROFILE_VAR("Nyx::reactions_alloc",var1);
     auto atomic_rates = atomic_rates_glob;
     auto f_rhs_data = (RhsData*)The_Arena()->alloc(sizeof(RhsData));
     Real gamma_minus_1 = gamma - 1.0;
@@ -157,15 +158,17 @@ int Nyx::integrate_state_struct_mfin
       {
                 if(sundials_alloc_type==0)
                   u = N_VMakeWithManagedAllocator_Cuda(neq,sunalloc,sunfree);  /* Allocate u vector */
-                else
+                else if(sundials_alloc_type==2)
                   u = N_VNewManaged_Cuda(neq);  /* Allocate u vector */
-		N_VSetCudaStream_Cuda(u, &currentStream);
+		else
+                  u = N_VNew_Cuda(neq);  /* Allocate u vector */
+                N_VSetCudaStream_Cuda(u, &currentStream);
                 amrex::Gpu::Device::streamSynchronize();
       }
       else
       {
-	if(sundials_alloc_type==1)
-	  {
+        if(sundials_alloc_type==1)
+          {
                 dptr=(realtype*) The_Arena()->alloc(neq*sizeof(realtype));
                 u = N_VMakeManaged_Cuda(neq,dptr);  /* Allocate u vector */
                 eptr= (realtype*) The_Arena()->alloc(neq*sizeof(realtype));
@@ -176,36 +179,41 @@ int Nyx::integrate_state_struct_mfin
                 abstol_ptr = (realtype*) The_Arena()->alloc(neq*sizeof(realtype));
                 abstol_vec = N_VMakeManaged_Cuda(neq,abstol_ptr);
                 N_VSetCudaStream_Cuda(abstol_vec,&currentStream);
+                T_vode=(realtype*) The_Arena()->alloc(neq*sizeof(realtype));
+                T_vec = N_VMakeManaged_Cuda(neq, T_vode);
+                ne_vode=(realtype*) The_Arena()->alloc(neq*sizeof(realtype));
+                ne_vec = N_VMakeManaged_Cuda(neq, ne_vode);
+                rho_vode=(realtype*) The_Arena()->alloc(neq*sizeof(realtype));
+                rho_vec = N_VMakeManaged_Cuda(neq, rho_vode);
+                rho_init_vode=(realtype*) The_Arena()->alloc(neq*sizeof(realtype));
+                rho_init_vec = N_VMakeManaged_Cuda(neq, rho_init_vode);
+                rho_src_vode=(realtype*) The_Arena()->alloc(neq*sizeof(realtype));
+                rho_src_vec = N_VMakeManaged_Cuda(neq, rho_src_vode);
+                rhoe_src_vode=(realtype*) The_Arena()->alloc(neq*sizeof(realtype));
+                rhoe_src_vec = N_VMakeManaged_Cuda(neq, rhoe_src_vode);
+                e_src_vode=(realtype*) The_Arena()->alloc(neq*sizeof(realtype));
+                e_src_vec = N_VMakeManaged_Cuda(neq, e_src_vode);
+                IR_vode=(realtype*) The_Arena()->alloc(neq*sizeof(realtype));
+                IR_vec = N_VMakeManaged_Cuda(neq, IR_vode);
[…]
