### 2021-04-30 2afdddf
Add Hip sundials allocation and fusing test
