+++ Source/HeatCool/integrate_state_with_source_3d.cpp
@@ -153,14 +153,14 @@ int Nyx::integrate_state_struct_mfin
       int loop = 1;
 
 #ifdef AMREX_USE_CUDA
-      cudaStream_t currentStream = amrex::Gpu::Device::cudaStream();  
+      cudaStream_t currentStream = amrex::Gpu::Device::cudaStream();
       if(sundials_alloc_type%2==0)
       {
                 if(sundials_alloc_type==0)
                   u = N_VMakeWithManagedAllocator_Cuda(neq,sunalloc,sunfree);  /* Allocate u vector */
                 else if(sundials_alloc_type==2)
                   u = N_VNewManaged_Cuda(neq);  /* Allocate u vector */
-		else
+                else
                   u = N_VNew_Cuda(neq);  /* Allocate u vector */
                 N_VSetCudaStream_Cuda(u, &currentStream);
                 amrex::Gpu::Device::streamSynchronize();
@@ -210,11 +210,11 @@ int Nyx::integrate_state_struct_mfin
             u = N_VNewWithMemHelp_Cuda(neq, /*use_managed_mem=*/true, *amrex::sundials::The_SUNMemory_Helper());
             N_VSetCudaStream_Cuda(u, &currentStream);
           }
-	else if(sundials_alloc_type==5)
+        else if(sundials_alloc_type==5)
           {
-	    u = N_VNewWithMemHelp_Cuda(neq, /*use_managed_mem=*/false, *amrex::sundials::The_SUNMemory_Helper());
+            u = N_VNewWithMemHelp_Cuda(neq, /*use_managed_mem=*/false, *amrex::sundials::The_SUNMemory_Helper());
             N_VSetCudaStream_Cuda(u, &currentStream);
-	  }
+          }
       }
 
       amrex::Gpu::streamSynchronize();
@@ -225,7 +225,46 @@ int Nyx::integrate_state_struct_mfin
 #else
 #ifdef AMREX_USE_GPU
 #ifdef AMREX_USE_HIP
-              u = N_VNew_Hip(neq);  /* Allocate u vector */
+      auto currentstream = amrex::Gpu::Device::gpuStream();
+      //Choosing 256 here since this mimics Sundials default
+      // Possibly attempt to match n_threads_and_blocks
+      //     long N = ((tbx.numPts()+AMREX_GPU_NCELLS_PER_THREAD-1)/AMREX_GPU_NCELLS_PER_THREAD);
+      //     SUNHipThreadDirectExecPolicy stream_exec_policy(AMREX_GPU_MAX_THREADS, currentstream);
+      //     SUNHipGridStrideExecPolicy grid_exec_policy(AMREX_GPU_MAX_THREADS, std::max((N + AMREX_GPU_MAX_THREADS - 1) / AMREX_GPU_MAX_THREADS, static_cast<Long>(1)), currentstream);
+      //     SUNHipBlockReduceExecPolicy reduce_exec_policy(AMREX_GPU_MAX_THREADS, std::max((N + AMREX_GPU_MAX_THREADS - 1) / AMREX_GPU_MAX_THREADS, static_cast<Long>(1)), currentstream);
+
+      SUNHipThreadDirectExecPolicy stream_exec_policy(256, currentstream);
+      SUNHipBlockReduceExecPolicy reduce_exec_policy(256, 0, currentstream);
+      if(sundials_alloc_type%2==0)
+      {
+                if(sundials_alloc_type==0)
+                  amrex::Abort("sundials_alloc_type=0 not implemented with Hip");
+                else if(sundials_alloc_type==2)
+                  u = N_VNewManaged_Hip(neq);  /* Allocate u vector */
+                else
+                  u = N_VNew_Hip(neq);  /* Allocate u vector */
[…]
