### 2021-05-04 78fdf6d
Tweak default sundials memory alloc flags
