+++ Source/HeatCool/integrate_state_with_source_3d.cpp
@@ -57,10 +57,10 @@ int Nyx::integrate_state_struct
   //#endif
 #ifdef _OPENMP
 #ifdef AMREX_USE_GPU
-    if (sundials_use_tiling)
+  //    if (sundials_use_tiling)
        const auto tiling = MFItInfo().SetDynamic(true);
-    else
-       const bool tiling = false;
+       //    else
+       //       const bool tiling = false;
 #pragma omp parallel
 #else
     const bool tiling = (TilingIfNotGPU() && sundials_use_tiling);
@@ -107,6 +107,7 @@ int Nyx::integrate_state_struct_mfin
    long int& old_max_steps, long int& new_max_steps,
    const int sdc_iter)
 {
+    BL_PROFILE_VAR("Nyx::reactions_alloc",var1);
     auto atomic_rates = atomic_rates_glob;
     auto f_rhs_data = (RhsData*)The_Arena()->alloc(sizeof(RhsData));
     Real gamma_minus_1 = gamma - 1.0;
@@ -138,6 +139,7 @@ int Nyx::integrate_state_struct_mfin
 
       void *cvode_mem;
       realtype *dptr, *eptr, *rpar, *rparh, *abstol_ptr;
+      Real *T_vode, *ne_vode,*rho_vode,*rho_init_vode,*rho_src_vode,*rhoe_src_vode,*e_src_vode,*IR_vode;
       realtype t=0.0;
                                 
       u = NULL;
@@ -153,15 +155,72 @@ int Nyx::integrate_state_struct_mfin
       amrex::Gpu::streamSynchronize();
       int loop = 1;
 
-      if(verbose>1)
-        amrex::Arena::PrintUsage();
 #ifdef AMREX_USE_CUDA
       cudaStream_t currentStream = amrex::Gpu::Device::cudaStream();
-      if(sundials_alloc_type==0)
-          u = N_VNewWithMemHelp_Cuda(neq, /*use_managed_mem=*/true, *amrex::sundials::The_SUNMemory_Helper());
+      if(sundials_alloc_type%2==0)
+      {
+                if(sundials_alloc_type==0)
+                  u = N_VMakeWithManagedAllocator_Cuda(neq,sunalloc,sunfree);  /* Allocate u vector */
+                else if(sundials_alloc_type==2)
+                  u = N_VNewManaged_Cuda(neq);  /* Allocate u vector */
+                else
+                  u = N_VNew_Cuda(neq);  /* Allocate u vector */
+                N_VSetCudaStream_Cuda(u, &currentStream);
+                amrex::Gpu::Device::streamSynchronize();
+      }
       else
-          u = N_VNewManaged_Cuda(neq);
-      N_VSetCudaStream_Cuda(u, &currentStream);
+      {
+        if(sundials_alloc_type==1)
+          {
+                dptr=(realtype*) The_Arena()->alloc(neq*sizeof(realtype));
+                u = N_VMakeManaged_Cuda(neq,dptr);  /* Allocate u vector */
+                eptr= (realtype*) The_Arena()->alloc(neq*sizeof(realtype));
+                e_orig = N_VMakeManaged_Cuda(neq,eptr);  /* Allocate u vector */
+                N_VSetCudaStream_Cuda(e_orig, &currentStream);
+                N_VSetCudaStream_Cuda(u, &currentStream);
+
+                abstol_ptr = (realtype*) The_Arena()->alloc(neq*sizeof(realtype));
+                abstol_vec = N_VMakeManaged_Cuda(neq,abstol_ptr);
+                N_VSetCudaStream_Cuda(abstol_vec,&currentStream);
[…]
