+++ .github/workflows/linux.yml
@@ -55,7 +55,7 @@ jobs:
             -DCMAKE_CUDA_HOST_COMPILER=$(which g++)          \
             -DAMReX_CUDA_ARCH=6.0
         cmake --build build_adiabatic --parallel 2
-	
+
         cmake  -S . -B build                                 \
             -DNyx_HEATCOOL=yes                               \
             -DNyx_HYDRO=yes                                  \
