+++ .github/workflows/dependencies/dependencies_nvcc.sh
@@ -15,7 +15,7 @@ sudo apt-get update
 
 sudo apt-get install -y --no-install-recommends\
     build-essential     \
-    g++-4.8             \
+    g++                 \
     libopenmpi-dev      \
     openmpi-bin         \
     nvidia-cuda-dev     \
