### 2021-05-05 9315261
Update Cuda tests in Github CI
