+++ .github/workflows/linux.yml
@@ -36,13 +36,14 @@ jobs:
             -DCMAKE_CXX_COMPILER=$(which clang++)
         cmake --build build --parallel 2
 
-  nyx-cuda:
-    name: CUDA@9.1.85 GNU@4.8.5 C++11 Release
-    runs-on: ubuntu-latest
+  nyx-cuda9:
+    name: CUDA@9.2 GNU@6.5.0 C++14 Release
+    runs-on: ubuntu-18.04
+    env: {CXXFLAGS: "-fno-operator-names"}
     steps:
     - uses: actions/checkout@v2
     - name: Dependencies
-      run: .github/workflows/dependencies/dependencies_nvcc.sh
+      run: .github/workflows/dependencies/dependencies_nvcc9.sh
     - name: Build & Install
       run: |
         cmake  -S . -B build_adiabatic                       \
@@ -71,17 +72,14 @@ jobs:
 #   https://gitlab.com/nvidia/container-images/cuda/-/blob/master/dist/ubuntu18.04/10.1/base/Dockerfile
 #   https://github.com/ComputationalRadiationPhysics/picongpu/blob/0.5.0/share/picongpu/dockerfiles/ubuntu-1604/Dockerfile
 #   https://developer.download.nvidia.com/compute/cuda/repos/ubuntu1804/x86_64/
-  build_nvcc:
+  nyx-cuda11:
     name: NVCC 11.0.2 SP [Linux]
     runs-on: ubuntu-20.04
     env: {CXXFLAGS: "-fno-operator-names"}
     steps:
     - uses: actions/checkout@v2
     - name: install dependencies
-      run: |
-        .github/workflows/dependencies/dependencies_nvcc11.sh
-        export CEI_SUDO="sudo"
-        cmake-easyinstall --prefix=/usr/local git+https://github.com/openPMD/openPMD-api.git@0.13.2 -DopenPMD_USE_PYTHON=OFF -DBUILD_TESTING=OFF -DBUILD_EXAMPLES=OFF -DBUILD_CLI_TOOLS=OFF
+      run: .github/workflows/dependencies/dependencies_nvcc11.sh
     - name: build Nyx
       run: |
         export PATH=/usr/local/nvidia/bin:/usr/local/cuda/bin:${PATH}
