### 2021-05-05 030b262
Update Cuda CI release targets and names
