### 2021-05-05 85177f0
Update Cuda CI flags and locations
