+++ .github/workflows/linux.yml
@@ -46,6 +46,8 @@ jobs:
       run: .github/workflows/dependencies/dependencies_nvcc9.sh
     - name: Build & Install
       run: |
+        echo $(which nvcc)
+	echo $(ls /usr/local/cuda/bin/)
         cmake  -S . -B build_adiabatic                       \
             -DNyx_HEATCOOL=no                                \
             -DNyx_HYDRO=yes                                  \
@@ -55,6 +57,7 @@ jobs:
             -DCMAKE_CXX_COMPILER=$(which g++-6)              \
             -DCMAKE_CUDA_HOST_COMPILER=$(which g++-6)        \
             -DCMAKE_CUDA_COMPILER=$(which nvcc)              \
+            -DCMAKE_CUDA_ARCHITECTURES=60                    \
             -DAMReX_CUDA_ARCH=6.0
         cmake --build build_adiabatic --parallel 2
 
@@ -67,6 +70,7 @@ jobs:
             -DCMAKE_CXX_COMPILER=$(which g++-6)              \
             -DCMAKE_CUDA_HOST_COMPILER=$(which g++-6)        \
             -DCMAKE_CUDA_COMPILER=$(which nvcc)              \
+            -DCMAKE_CUDA_ARCHITECTURES=60                    \
             -DAMReX_CUDA_ARCH=6.0
         cmake --build build --parallel 2
 
@@ -97,6 +101,7 @@ jobs:
             -DCMAKE_CUDA_HOST_COMPILER=$(which g++)          \
             -DCMAKE_CUDA_STANDARD=17                         \
             -DCMAKE_CXX_STANDARD=17                          \
+            -DCMAKE_CUDA_ARCHITECTURES=80                    \
             -DAMReX_CUDA_ARCH=8.0
         cmake --build build_adiabatic --parallel 2
 
@@ -110,6 +115,7 @@ jobs:
             -DCMAKE_CUDA_HOST_COMPILER=$(which g++)          \
             -DCMAKE_CUDA_STANDARD=17                         \
             -DCMAKE_CXX_STANDARD=17                          \
+            -DCMAKE_CUDA_ARCHITECTURES=80                    \
             -DAMReX_CUDA_ARCH=8.0
         cmake --build build --parallel 2
 
