### 2021-05-05 9c5ce3f
Add cusparse to CI and remove noexcept from ReduceMin
