+++ .github/workflows/dependencies/dependencies_nvcc11.sh
@@ -36,5 +36,7 @@ sudo apt-get install -y \
     cuda-minimal-build-11-2      \
     cuda-nvml-dev-11-2           \
     cuda-nvtx-11-2               \
-    libcurand-dev-11-2
+    libcurand-dev-11-2           \
+    libcusparse-dev-11-2
+
 sudo ln -s cuda-11.2 /usr/local/cuda
