+++ .github/workflows/linux.yml
@@ -46,6 +46,8 @@ jobs:
       run: .github/workflows/dependencies/dependencies_nvcc9.sh
     - name: Build & Install
       run: |
+        export PATH=$PATH:/usr/local/cuda/bin
+        export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/local/cuda/lib
         cmake  -S . -B build_adiabatic                       \
             -DNyx_HEATCOOL=no                                \
             -DNyx_HYDRO=yes                                  \
@@ -54,7 +56,6 @@ jobs:
             -DCMAKE_C_COMPILER=$(which gcc-6)                \
             -DCMAKE_CXX_COMPILER=$(which g++-6)              \
             -DCMAKE_CUDA_HOST_COMPILER=$(which g++-6)        \
-            -DCMAKE_CUDA_COMPILER=/usr/local/cuda/bin/nvcc   \
             -DCMAKE_CUDA_ARCHITECTURES=60                    \
             -DAMReX_CUDA_ARCH=6.0
         cmake --build build_adiabatic --parallel 2
@@ -67,7 +68,6 @@ jobs:
             -DCMAKE_C_COMPILER=$(which gcc-6)                \
             -DCMAKE_CXX_COMPILER=$(which g++-6)              \
             -DCMAKE_CUDA_HOST_COMPILER=$(which g++-6)        \
-            -DCMAKE_CUDA_COMPILER=/usr/local/cuda/bin/nvcc   \
             -DCMAKE_CUDA_ARCHITECTURES=60                    \
             -DAMReX_CUDA_ARCH=6.0
         cmake --build build --parallel 2
