### 2021-05-05 d7a1aa4
Fix argument order for Sycl and first pass at cmake update
