+++ Source/HeatCool/integrate_state_with_source_3d.cpp
@@ -274,12 +274,13 @@ int Nyx::integrate_state_struct_mfin
       //Choosing 256 here since this mimics Sundials default
       // Possibly attempt to match n_threads_and_blocks
       //     long N = ((tbx.numPts()+AMREX_GPU_NCELLS_PER_THREAD-1)/AMREX_GPU_NCELLS_PER_THREAD);
-      //     SUNSyclThreadDirectExecPolicy stream_exec_policy(AMREX_GPU_MAX_THREADS, currentstream);
-      //     SUNSyclGridStrideExecPolicy grid_exec_policy(AMREX_GPU_MAX_THREADS, std::max((N + AMREX_GPU_MAX_THREADS - 1) / AMREX_GPU_MAX_THREADS, static_cast<Long>(1)), currentstream);
-      //     SUNSyclBlockReduceExecPolicy reduce_exec_policy(AMREX_GPU_MAX_THREADS, std::max((N + AMREX_GPU_MAX_THREADS - 1) / AMREX_GPU_MAX_THREADS, static_cast<Long>(1)), currentstream);
+      //     SUNSyclThreadDirectExecPolicy stream_exec_policy(AMREX_GPU_MAX_THREADS);
+      //     SUNSyclGridStrideExecPolicy grid_exec_policy(AMREX_GPU_MAX_THREADS, std::max((N + AMREX_GPU_MAX_THREADS - 1) / AMREX_GPU_MAX_THREADS, static_cast<Long>(1)));
+      //     SUNSyclBlockReduceExecPolicy reduce_exec_policy(AMREX_GPU_MAX_THREADS, std::max((N + AMREX_GPU_MAX_THREADS - 1) / AMREX_GPU_MAX_THREADS, static_cast<Long>(1)));
 
-      SUNSyclThreadDirectExecPolicy stream_exec_policy(256, currentstream);
-      SUNSyclBlockReduceExecPolicy reduce_exec_policy(256, 0, currentstream);
+      //Sycl version does not take a stream or queue
+      SUNSyclThreadDirectExecPolicy stream_exec_policy(256);
+      SUNSyclBlockReduceExecPolicy reduce_exec_policy(256, 0);
       if(sundials_alloc_type%2==0)
       {
                 if(sundials_alloc_type==0)
@@ -310,16 +311,16 @@ int Nyx::integrate_state_struct_mfin
         else if(sundials_alloc_type==7)
         {
               //   Directly providing MemHelp:
-              e_orig = N_VNewWithMemHelp_Sycl(neq, &amrex::Gpu::Device::streamQueue(), /*use_managed_mem=*/false, *amrex::sundials::The_SUNMemory_Helper());
-              abstol_vec = N_VNewWithMemHelp_Sycl(neq, &amrex::Gpu::Device::streamQueue(), /*use_managed_mem=*/false, *amrex::sundials::The_SUNMemory_Helper());
-              T_vec = N_VNewWithMemHelp_Sycl(neq, &amrex::Gpu::Device::streamQueue(), /*use_managed_mem=*/false, *amrex::sundials::The_SUNMemory_Helper());
-              ne_vec = N_VNewWithMemHelp_Sycl(neq, &amrex::Gpu::Device::streamQueue(), /*use_managed_mem=*/false, *amrex::sundials::The_SUNMemory_Helper());
-              rho_vec = N_VNewWithMemHelp_Sycl(neq, &amrex::Gpu::Device::streamQueue(), /*use_managed_mem=*/false, *amrex::sundials::The_SUNMemory_Helper());
[…]
