### 2021-05-05 f8c2fde
Update default DPCPP Sundials alloc
