### 2021-05-06 0b2b94e
Make append option for dark matter hdf5
