### 2021-05-07 bf60fd5
Fix example script
