### 2021-05-07 e420d8e
Add print statements to CI
