### 2021-05-07 1635bf9
Add cusolver for Sundials CI
