+++ .github/workflows/dependencies/dependencies_nvcc11.sh
@@ -37,6 +37,7 @@ sudo apt-get install -y \
     cuda-nvml-dev-11-2           \
     cuda-nvtx-11-2               \
     libcurand-dev-11-2           \
+    libcusolver-dev-11-2           \
     libcusparse-dev-11-2
 
 sudo ln -s cuda-11.2 /usr/local/cuda
