+++ .github/workflows/linux.yml
@@ -46,8 +46,8 @@ jobs:
       run: .github/workflows/dependencies/dependencies_nvcc9.sh
     - name: Build & Install
       run: |
-        export PATH=$PATH:/usr/local/cuda/bin
-        export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/local/cuda/lib
+        export PATH=/usr/local/nvidia/bin:/usr/local/cuda/bin:${PATH}
+        export LD_LIBRARY_PATH=/usr/local/nvidia/lib:/usr/local/nvidia/lib64:/usr/local/cuda/lib64:/usr/local/cuda/lib:${LD_LIBRARY_PATH}
         ls /usr/local/cuda/*
         cmake  -S . -B build_adiabatic                       \
             -DNyx_HEATCOOL=no                                \
@@ -89,10 +89,9 @@ jobs:
     - name: build Nyx
       run: |
         export PATH=/usr/local/nvidia/bin:/usr/local/cuda/bin:${PATH}
-        export LD_LIBRARY_PATH=/usr/local/nvidia/lib:/usr/local/nvidia/lib64:/usr/local/cuda/lib64:${LD_LIBRARY_PATH}
-        ls /usr/local/cuda/*
-        ls /usr/local/nvidia/*
+        export LD_LIBRARY_PATH=/usr/local/nvidia/lib:/usr/local/nvidia/lib64:/usr/local/cuda/lib64:/usr/local/cuda/lib:${LD_LIBRARY_PATH}
         which nvcc || echo "nvcc not in PATH!"
+        ls /usr/local/cuda/*
         cmake  -S . -B build_adiabatic                       \
             -DNyx_HEATCOOL=no                                \
             -DNyx_HYDRO=yes                                  \
@@ -112,6 +111,7 @@ jobs:
             -DNyx_HYDRO=yes                                  \
             -DNyx_OMP=no                                     \
             -DNyx_GPU_BACKEND=CUDA                           \
+            -DCMAKE_VERBOSE_MAKEFILE=ON                      \
             -DCMAKE_C_COMPILER=$(which gcc)                  \
             -DCMAKE_CXX_COMPILER=$(which g++)                \
             -DCMAKE_CUDA_HOST_COMPILER=$(which g++)          \
