### 2021-05-21 88c22e4
Second pass adding hip with omp
