+++ Source/EOS/atomic_rates.H
@@ -39,25 +39,25 @@ void tabulate_rates(std::string file_in, amrex::Real mean_rhob)
         for(i = 1; i <= NCOOLTAB+1; i++)
         {
             // Rates are as in Katz et al. 1996
-            sqrt_t = std::sqrt(t);
-            corr_term    = 1.e0 / (1.0e0 + sqrt_t/std::sqrt(t5));
+            sqrt_t = ::sqrt(t);
+            corr_term    = 1.e0 / (1.0e0 + sqrt_t/::sqrt(t5));
 
             // Recombination rates
             // Alphad: dielectronic recombination rate of singly ioniozed helium
-            atomic_rates_host.Alphad(i)    = 1.90e-03/(t*sqrt_t) * std::exp(-4.7e5/t) * (1.0e0+0.3e0*std::exp(-9.4e4/t));
+            atomic_rates_host.Alphad(i)    = 1.90e-03/(t*sqrt_t) * ::exp(-4.7e5/t) * (1.0e0+0.3e0*::exp(-9.4e4/t));
             atomic_rates_host.AlphaHp(i)   = 8.40e-11/sqrt_t * std::pow((t/t3),(-0.2e0)) / (1.0e0 + std::pow((t/t6),0.7e0));
             atomic_rates_host.AlphaHep(i)  = 1.50e-10 * std::pow(t,(-0.6353e0));
             atomic_rates_host.AlphaHepp(i) = 3.36e-10/sqrt_t * std::pow((t/t3),(-0.2e0)) / (1.0e0 + std::pow((t/t6),0.7e0));
 
             // Collisional ionization rates
-            atomic_rates_host.GammaeH0(i)  = 5.85e-11*sqrt_t * std::exp(-157809.1e0/t) * corr_term;
-            atomic_rates_host.GammaeHe0(i) = 2.38e-11*sqrt_t * std::exp(-285335.4e0/t) * corr_term;
-            atomic_rates_host.GammaeHep(i) = 5.68e-12*sqrt_t * std::exp(-631515.0e0/t) * corr_term;
+            atomic_rates_host.GammaeH0(i)  = 5.85e-11*sqrt_t * ::exp(-157809.1e0/t) * corr_term;
+            atomic_rates_host.GammaeHe0(i) = 2.38e-11*sqrt_t * ::exp(-285335.4e0/t) * corr_term;
+            atomic_rates_host.GammaeHep(i) = 5.68e-12*sqrt_t * ::exp(-631515.0e0/t) * corr_term;
 
             // Collisional ionization & excitation cooling rates
-            atomic_rates_host.BetaH0(i)  = 7.5e-19 * std::exp(-118348.0e0/t) * corr_term + 2.171e-11*atomic_rates_host.GammaeH0(i);
+            atomic_rates_host.BetaH0(i)  = 7.5e-19 * ::exp(-118348.0e0/t) * corr_term + 2.171e-11*atomic_rates_host.GammaeH0(i);
             atomic_rates_host.BetaHe0(i) = 3.941e-11 * atomic_rates_host.GammaeHe0(i);
-            atomic_rates_host.BetaHep(i) = 5.54e-17 * std::pow(t,(-0.397e0)) * std::exp(-473638.0e0/t) * corr_term + 0.715e-11 * atomic_rates_host.GammaeHep(i);
+            atomic_rates_host.BetaHep(i) = 5.54e-17 * std::pow(t,(-0.397e0)) * ::exp(-473638.0e0/t) * corr_term + 0.715e-11 * atomic_rates_host.GammaeHep(i);
 
             // Recombination cooling rates
             atomic_rates_host.RecHp(i)   = 1.036e-16 * t * atomic_rates_host.AlphaHp(i);
[…]
