### 2021-05-25 #3068 @teuben
yes, i had downloaded the code, and browsed around it's capabilities. I noted (happily) that Nyx is already in ascl on https://ascl.net/1712.006, so it's been around  a while.
But I also noted it's been 3 weeks, so easy to slip in time.  I'll get some hours in this week, it's a fun code.
