### 2021-05-26 32162ac
Add back in instantaneous heating from reionization
