+++ Source/HeatCool/f_rhs_struct.H
@@ -15,7 +15,7 @@ struct RhsData {
     int         flash_he;
     amrex::Real zhi_flash;
     amrex::Real zheii_flash;
-    amrex::Real T_zhii;
+    amrex::Real T_zhi;
     amrex::Real T_zheii;
     amrex::Real H_reion_z;
     amrex::Real He_reion_z;
@@ -56,7 +56,7 @@ ode_eos_setup(RhsData* f_rhs_data, amrex::Real gamma_minus_1, amrex::Real h_spec
     f_rhs_data_host.uvb_density_B=0.0;
     f_rhs_data_host.zhi_flash=-1.0;
     f_rhs_data_host.zheii_flash=-1.0;
-    f_rhs_data_host.T_zhii=0.0;
+    f_rhs_data_host.T_zhi=0.0;
     f_rhs_data_host.T_zheii=0.0;
     f_rhs_data_host.flash_h=0;
     f_rhs_data_host.flash_he=0;
@@ -65,7 +65,7 @@ ode_eos_setup(RhsData* f_rhs_data, amrex::Real gamma_minus_1, amrex::Real h_spec
     pp_nyx.query("uvb_density_B"            , f_rhs_data_host.uvb_density_B);
     pp_nyx.query("reionization_zHI_flash"   , f_rhs_data_host.zhi_flash);
     pp_nyx.query("reionization_zHeII_flash" , f_rhs_data_host.zheii_flash);
-    pp_nyx.query("reionization_T_zHI"       , f_rhs_data_host.T_zhii);
+    pp_nyx.query("reionization_T_zHI"       , f_rhs_data_host.T_zhi);
     pp_nyx.query("reionization_T_zHeII"     , f_rhs_data_host.T_zheii);
 
     // Set options in reion_aux_module                                                                                                                                                                                                        
@@ -276,16 +276,88 @@ ode_eos_finalize_struct(const int i, const int j, const int k, const int idx, At
         e_out  = f_rhs_data->T_vode[idx] / (f_rhs_data->gamma_minus_1 * mp_over_kb * mu);
     }
     }
+
+    Real nh0, nhep, nhp, nhe0, nhepp;
     /////////////////   //update nyx_eos_T_given_Re
     if (f_rhs_data->inhomogeneous_on)
-       nyx_eos_T_given_Re_device(atomic_rates, f_rhs_data->gamma_minus_1, f_rhs_data->h_species, f_rhs_data->JH_vode_arr[idx], f_rhs_data->JHe_vode, &(f_rhs_data->T_vode[idx]), &(f_rhs_data->ne_vode[idx]), f_rhs_data->rho_vode[idx], e_out, f_rhs_data->a);
+      nyx_eos_T_given_Re_device(atomic_rates, f_rhs_data->gamma_minus_1, f_rhs_data->h_species, f_rhs_data->JH_vode_arr[idx], f_rhs_data->JHe_vode, &(f_rhs_data->T_vode[idx]), &(f_rhs_data->ne_vode[idx]), f_rhs_data->rho_vode[idx], e_out, f_rhs_data->a, nh0, nhep, nhp, nhe0, nhepp);
     else
[…]
