### 2021-05-26 8e2812c
Add compile flag to exclude SUNMemoryHelper
