+++ Source/HeatCool/integrate_state_with_source_3d.cpp
@@ -1,7 +1,9 @@
 #include <fstream>
 #include <iomanip>
 
+#ifdef AMREX_USE_SUNDIALS_SUNMEMORY
 #include <AMReX_SUNMemory.H>
+#endif
 #include <cvode/cvode.h>               /* prototypes for CVODE fcts., consts. */
 #include <cvode/cvode_diag.h>          /* access to CVDiag interface */
 #include <sundials/sundials_types.h>   /* definition of type realtype */
@@ -148,8 +150,10 @@ int Nyx::integrate_state_struct_mfin
       abstol_vec = NULL;
       cvode_mem = NULL;
 
+#ifdef AMREX_USE_SUNDIALS_SUNMEMORY
 #ifdef AMREX_USE_DPCPP
       SUNMemoryHelper S = SUNMemoryHelper_Sycl(&amrex::Gpu::Device::streamQueue());
+#endif
 #endif
       long int neq = len.x*len.y*len.z;
       amrex::Gpu::streamSynchronize();
@@ -208,6 +212,7 @@ int Nyx::integrate_state_struct_mfin
                 N_VSetCudaStream_Cuda(IR_vec,&currentStream);
                 amrex::Gpu::streamSynchronize();
           }
+#ifdef AMREX_USE_SUNDIALS_SUNMEMORY
         else if(sundials_alloc_type==3)
           {
             u = N_VNewWithMemHelp_Cuda(neq, /*use_managed_mem=*/true, *amrex::sundials::The_SUNMemory_Helper());
@@ -218,6 +223,7 @@ int Nyx::integrate_state_struct_mfin
             u = N_VNewWithMemHelp_Cuda(neq, /*use_managed_mem=*/false, *amrex::sundials::The_SUNMemory_Helper());
             N_VSetCudaStream_Cuda(u, &currentStream);
           }
+#endif
       }
 
       amrex::Gpu::streamSynchronize();
@@ -255,6 +261,7 @@ int Nyx::integrate_state_struct_mfin
           {
                   amrex::Abort("sundials_alloc_type=1 not implemented with Hip");
           }
+#ifdef AMREX_USE_SUNDIALS_SUNMEMORY
         else if(sundials_alloc_type==3)
           {
             u = N_VNewWithMemHelp_Hip(neq, /*use_managed_mem=*/true, *amrex::sundials::The_SUNMemory_Helper());
@@ -263,6 +270,7 @@ int Nyx::integrate_state_struct_mfin
           {
             u = N_VNewWithMemHelp_Hip(neq, /*use_managed_mem=*/false, *amrex::sundials::The_SUNMemory_Helper());
           }
+#endif
 // might need a cuda analog to setting exec policy
       }
       N_VSetKernelExecPolicy_Hip(u, &stream_exec_policy, &reduce_exec_policy);
@@ -298,6 +306,7 @@ int Nyx::integrate_state_struct_mfin
           {
                   amrex::Abort("sundials_alloc_type=1 not implemented with Sycl");
           }
+#ifdef AMREX_USE_SUNDIALS_SUNMEMORY
         else if(sundials_alloc_type==3)
           {
               u = N_VNewWithMemHelp_Sycl(neq, /*use_managed_mem=*/true, *amrex::sundials::The_SUNMemory_Helper(), &amrex::Gpu::Device::streamQueue());
@@ -308,15 +317,10 @@ int Nyx::integrate_state_struct_mfin
               u = N_VNewWithMemHelp_Sycl(neq, /*use_managed_mem=*/false, *amrex::sundials::The_SUNMemory_Helper(), &amrex::Gpu::Device::streamQueue());
[…]
