### 2021-05-27 96f2b46
Change submodule to development. Make default sundials_alloc_type nonempty
