### 2021-05-27 0aec1b3
Update override and deprecated copy to fix warnings
