### 2021-05-27 c03af74
Add rocprim to HIP CI
