+++ .github/workflows/dependencies/dependencies_hip.sh
@@ -33,7 +33,10 @@ sudo apt-get install -y --no-install-recommends \
     libnuma-dev     \
     libopenmpi-dev  \
     openmpi-bin     \
-    rocm-dev rocrand
+    rocm-dev        \
+    rocfft          \
+    rocprim         \
+    rocrand
 
 # activate
 #
