### 2021-05-27 0d7e0c7
Rearrange ifdefs for clarity on GPU type
