+++ cmake/NyxSetupAMReX.cmake
@@ -107,6 +107,7 @@ else ()
    set(AMReX_PARTICLES            ON                         CACHE INTERNAL "" )
    set(AMReX_SUNDIALS             ${Nyx_SUNDIALS}            CACHE INTERNAL "" )
    set(AMReX_BUILD_TUTORIALS      OFF                        CACHE INTERNAL "" )
+   set(AMReX_INSTALL              OFF                        CACHE INTERNAL "" )
    if (Nyx_SINGLE_PRECISION_PARTICLES)
       set(AMReX_PARTICLES_PRECISION SINGLE CACHE INTERNAL "" )
    else ()
