### 2021-05-27 0985311
Add CMake compile option for SunMemory
