+++ cmake/NyxSetupAMReX.cmake
@@ -105,7 +105,7 @@ else ()
    set(AMReX_GPU_BACKEND          ${Nyx_GPU_BACKEND}         CACHE INTERNAL "" )
    set(AMReX_LINEAR_SOLVERS       ${Nyx_GRAVITY}             CACHE INTERNAL "" )
    set(AMReX_PARTICLES            ON                         CACHE INTERNAL "" )
-   set(AMReX_SUNDIALS             ${Nyx_SUNDIALS}            CACHE INTERNAL "" )
+   set(AMReX_SUNDIALS             ${Nyx_SUNDIALS_SUNMEMORY}  CACHE INTERNAL "" )
    set(AMReX_BUILD_TUTORIALS      OFF                        CACHE INTERNAL "" )
    set(AMReX_INSTALL              OFF                        CACHE INTERNAL "" )
    if (Nyx_SINGLE_PRECISION_PARTICLES)
