### 2021-05-28 8dd3a67
Update documentation for runs with Sundials and Hip
