### 2021-05-28 f2b0a37
GPU and OpenMP picky about sundials includes order
Fix typo in make Sunmemory compile guard
