+++ Source/HeatCool/integrate_state_with_source_3d.cpp
@@ -1,21 +1,7 @@
 #include <fstream>
 #include <iomanip>
 
-#ifdef AMREX_USE_SUNDIALS_SUNMEMORY
-#include <AMReX_SUNMemory.H>
-#endif
-
-#include <AMReX_ParmParse.H>
-#include <AMReX_Geometry.H>
-#include <AMReX_MultiFab.H>
-#include <AMReX_Print.H>
-#include <AMReX_PlotFileUtil.H>
-
-#include <AMReX_BLFort.H>
-#include <Nyx.H>
-#include <f_rhs.H>
-#include <f_rhs_struct.H>
-
+#ifndef AMREX_USE_DPCPP
 #include <cvode/cvode.h>               /* prototypes for CVODE fcts., consts. */
 #include <cvode/cvode_diag.h>          /* access to CVDiag interface */
 #include <sundials/sundials_types.h>   /* definition of type realtype */
@@ -32,6 +18,42 @@
 #endif
 #ifdef AMREX_USE_DPCPP
 #include <nvector/nvector_sycl.h>
+#endif
+#endif
+
+#ifdef AMREX_USE_SUNDIALS_SUNMEMORY
+#include <AMReX_SUNMemory.H>
+#endif
+
+#include <AMReX_ParmParse.H>
+#include <AMReX_Geometry.H>
+#include <AMReX_MultiFab.H>
+#include <AMReX_Print.H>
+#include <AMReX_PlotFileUtil.H>
+
+#include <AMReX_BLFort.H>
+#include <Nyx.H>
+#include <f_rhs.H>
+#include <f_rhs_struct.H>
+
+#ifdef AMREX_USE_DPCPP
+#include <cvode/cvode.h>               /* prototypes for CVODE fcts., consts. */
+#include <cvode/cvode_diag.h>          /* access to CVDiag interface */
+#include <sundials/sundials_types.h>   /* definition of type realtype */
+
+#include <nvector/nvector_serial.h>
+#ifdef _OPENMP
+#include <nvector/nvector_openmp.h>
+#endif
+#ifdef AMREX_USE_CUDA
+#include <nvector/nvector_cuda.h>
+#endif
+#ifdef AMREX_USE_HIP
+#include <nvector/nvector_hip.h>
+#endif
+#ifdef AMREX_USE_DPCPP
+#include <nvector/nvector_sycl.h>
+#endif
 #endif
 
 //#define MAKE_MANAGED 1
