### 2021-06-02 a2c6756
derive plotfile use new jobinfo, the_parameters
This makes the plotfile to hdf5 converter aware of
universe metadata like comoving_h seperately from other flags,
and writes the _gridDM plotfile based on the newest inputs
