+++ Docs/sphinx_documentation/source/conf.py
@@ -117,16 +117,16 @@ todo_include_todos = True
 
 # -- Options for MathJax
 #mathjax_path = 'MathJax/MathJax.js'
-mathjax_config = {'TeX': {'Macros': {}}}
+mathjax3_config = {'tex': {'macros': {}}}
 
 with open('mathsymbols.tex', 'r') as f:
     for line in f:
         macros = re.findall(r'\\newcommand{\\(.*?)}(\[(\d)\])?{(.+)}', line)
         for macro in macros:
             if len(macro[1]) == 0:
-                mathjax_config['TeX']['Macros'][macro[0]] = "{"+macro[3]+"}"
+                mathjax3_config['tex']['macros'][macro[0]] = "{"+macro[3]+"}"
             else:
-                mathjax_config['TeX']['Macros'][macro[0]] = ["{"+macro[3]+"}", int(macro[2])]
+                mathjax3_config['tex']['macros'][macro[0]] = ["{"+macro[3]+"}", int(macro[2])]
 
 # -- Options for HTML output ----------------------------------------------
 
