### 2021-06-02 e3d9248
Move plotfile write in order to use write_particle_density_at_init and call Nyx functions directly
