+++ Util/Converters/assign_particle_vels/nyx_main.cpp
@@ -103,166 +103,6 @@ nyx_main (int argc, char* argv[])
     Amr *amrptr = new Amr(getLevelBld());
     amrptr->init(strt_time,stop_time);
 
-    // Regrid for now
-    //    amrptr->RegridOnly(amrptr->cumTime());
-
-    ParmParse pp_nyx("nyx");
-    std::string restart_particle_file;
-
-    pp_nyx.query("restart_particle_file",restart_particle_file);
-    //    Amr *amrptr = new Amr(getLevelBld());
-    /*
-    DarkMatterParticleContainer* DMPC = 0;
-    DMPC = new DarkMatterParticleContainer(amrptr);
-    DMPC->Restart(particle_plot_file, dm_chk_particle_file);
-    */
-    ////    Nyx::theDMPC()->Restart(particle_plot_file, dm_chk_particle_file);
-
-    auto dir = restart_particle_file + "_gridDM";
-
-    /*
-#ifdef NO_HYDRO
-    const Real cur_time = ((Nyx*) amrptr)->state[PhiGrav_Type].curTime();
-#else
-    const Real cur_time = ((Nyx*) amrptr)->state[State_Type].curTime();
-#endif
-*/
-    /*
-    const Real cur_time = ((Nyx*) amrptr)->initial_time;
-    int cycle = ((Nyx*) amrptr)->nStep();*/
-    const Real cur_time = 0.0;
-    int cycle = 0;
-
-    Vector<std::string> varnames;
-    const int n_data_items = 4;
-    const int nGrow = 1;
-    const int lev = 0;
-    MultiFab plotMF(Nyx::theDMPC()->ParticleBoxArray(lev),
-                    Nyx::theDMPC()->ParticleDistributionMap(lev),
-                    n_data_items,
-                    nGrow);
-    varnames.push_back("particle_mass_density");
-    varnames.push_back("particle_x_velocity");
-    varnames.push_back("particle_y_velocity");
-    varnames.push_back("particle_z_velocity");
-    //    varnames.push_back("particle_count");
-
-    Nyx::theDMPC()->AssignDensitySingleLevel(plotMF,lev,4);
-    /*
-    for(int cnt = 0; cnt < n_data_items; cnt++)
-    {
-        amrex::Print()<<"Getting derive type for: "<<varnames[cnt]<<std::endl;
-        amrex::Print()<<"with boxes and dmap "<<
-          Nyx::theDMPC()->ParticleBoxArray(lev)<<"\n"<<
-          Nyx::theDMPC()->ParticleDistributionMap(lev)<<"\n"
-          //<<    ((Nyx*) amrptr)->grids<<((Nyx*) amrptr)->dmap
-          <<std::endl;
-        const auto& derive_dat = ((Nyx*) amrptr)->particle_derive(varnames[cnt], cur_time, nGrow);
-        amrex::Print()<<"made derive"<<std::endl;
-        MultiFab::Copy(plotMF, *derive_dat, 0, cnt, 1, nGrow);
-    }*/
-    WriteSingleLevelPlotfile(dir,
-                             plotMF,
-                             varnames,
-                             Nyx::theDMPC()->GetParGDB()->Geom(lev),
-                             cur_time,
-                             cycle);
-
-    if (ParallelDescriptor::IOProcessor())
-      {
-          //Copy comoving_a file
-            std::string FileName = restart_particle_file + "/comoving_a";
-            std::ifstream FileIn;
-            std::string line;
-            FileIn.open(FileName.c_str(), std::ios::in);
-            if ( ! FileIn.good()) {
-                amrex::FileOpenFailed(FileName);
-            }
-            FileName = dir + "/comoving_a";
-            std::ofstream File;
[…]
