### 2021-06-07 #3068 @teuben
How strict are we only filenames.    If so, the license.txt file should be renamed to LICENSE
