### 2021-06-07 #3068 @dfm
Thanks @teuben! `license.txt` is an acceptable name for the license, although it looks like perhaps the copyright info is a little out of date. It would be fine (encouraged?) to change that to `LICENSE.txt` or `LICENSE` to be more consistent with standard practices, but I wouldn't require it!
