### 2021-06-07 #3068 @teuben
@jmsexton03 @dfm   I also went over the install using the documentation. It works, but I find it a bit frustrating that there are two install methods:   make and cmake.  with make there is no configure, plus the name of the executable (in e.g. the MiniSB example) depends not only on make/cmake, but even on the settings in the GNUMakefile. Thus I could not run an example via copy/paste. Is it really needed that the name of the executable contain all the provenance?   Let the user figure out a way if they really want to maintain cross-compile like executables. This would also simplify your documentation :-)   Just a thought.
