### 2021-06-07 #3068 @dfm
> @jmsexton03 @dfm I have some suggestion for the article. Here or in an issue?

@teuben: The standard procedure here would be to open an issue (or several) on AMReX-Astro/Nyx with your comments. It's useful (for me, especially!) if you can reference this issue thread on those issues.
