### 2021-06-07 AMReX-Astro/Nyx#80 @teuben
some paper.md comments
A few comments on the paper:

- Strang splitting :   maybe a  reference here ?  It's not my field, but I've not heard of this.
- six species:    I counted five
- "using geometric" -> "using the/a geometric" ?  (an article seems to be missing,
  but is it "the" or one of many ("a") ?
- "is necessity"  -> "is a necessity", i believe an article is missing
- "we do best effort" -> "we made a best effort" ( think,  I'm not a native speaker though)
- there ought to be a comment for poor MacOSX users here, since Linux is mentioned. Should they not bother?

From https://github.com/openjournals/joss-reviews/issues/3068
