### 2021-06-08 4d4c224
Small changes to the paper
