+++ Docs/paper/paper.bib
@@ -1,3 +1,15 @@
+@ARTICLE{strang1968,
+       author = {{Strang}, Gilbert},
+        title = "{On the Construction and Comparison of Difference Schemes}",
+      journal = {SIAM Journal on Numerical Analysis},
+         year = 1968,
+        month = sep,
+       volume = {5},
+       number = {3},
+        pages = {506-517},
+          doi = {10.1137/0705041}
+}
+
 @ARTICLE{onorbe2019,
     author = {{O{\~n}orbe}, Jose and {Davies}, F.~B. and {Luki{\'c}} and {}, Z. and {Hennawi}, J.~F. and {Sorini}, D.},
     title = "{Inhomogeneous reionization models in cosmological hydrodynamical simulations}",
