### 2021-06-08 boxlib-codes/nyx#80 @zarija
- Added a reference for Strang splitting.
- Six species are: neutral and ionized hydrogen, neutral, once and twice ionized helium, and free electrons.
- Made suggested grammar changes; if a native speaker can additionally go through it and fix other a/the articles that would be great.
- Added a brief comment about Macs...  Basically, it boils down to the fact that Nyx is really a high-performance computing (HPC) code, and there's no HPC platform running iOS.  Thus while Mac can be a convenient tool for Nyx code development, meaningful simulations cannot really be produced on Macs.
