+++ Docs/paper/paper.md
@@ -41,7 +41,7 @@ bibliography: paper.bib
 ``Nyx`` is a highly parallel, adaptive mesh, finite-volume 
 N-body compressible hydrodynamics solver for cosmological simulations.  
 It has been used to simulate different cosmological scenarios with
-a recent focus on the integalactic medium and Lyman alpha forest.
+a recent focus on the intergalactic medium and Lyman alpha forest.
 Together, Nyx, the compressible astrophysical simulation code, Castro [@castro], 
 and the low Mach number code MAESTROeX [@maestroex], make up the
 AMReX-Astrophysics Suite of open-source, adaptive mesh, performance-portable 
@@ -52,13 +52,13 @@ directionally unsplit corner transport upwind method of @ctu with
 piecewise parabolic reconstruction [@ppm].  In Nyx, we have
 several modes of coupling the stiff heating-cooling source terms to the hydro.  
 The simplest method is the traditional operator splitting approach, 
-using Strang splitting [@strang1968] to achieve second-order in time.  However, 
+using Strang splitting [@strang1968] to achieve second-order accuracy in time. However,
 this coupling can break down, and we have an alternative to Strang splitting
 based on spectral deferred corrections (SDC), a method
 that aims to prevent the hydro and stiff source terms from becoming decoupled.  
 The simplified SDC method uses the CTU PPM hydro together with an
 iterative scheme to fully couple the source terms and hydro, still to
-second order [@simple_sdc].
+second-order accuracy in time [@simple_sdc].
 
 Nyx has a set of additional physics necessary to model the intergalactic medium
 using heating-cooling source terms.
@@ -69,7 +69,7 @@ modeled in the code. Heating and cooling source terms are calculated using
 a sub-cycled approach in order to avoid running the whole code on a short, cooling timescale.
 Cosmological reionization is accounted for via a spatially uniform,
 but time-varying ultraviolet background (UVB) radiation field,
-inputed to the code as a list of photoionization and photoheating
+inputted to the code as a list of photoionization and photoheating
 rates that vary with redshift. Nyx also has the capability to model flash reionization
 as well as inhomogeneous reionization [@onorbe2019].
 
@@ -83,7 +83,7 @@ model neutrino particle effects and active galactic nuclei feedback.
 
 Nyx is built on the AMReX [@AMReX] adaptive mesh refinement (AMR)
 library and is written in C++.
-AMR levels are advanced at their own timestep (subcycling)
+AMR levels are advanced at their own timestep (sub-cycling)
 and jumps by factors of 2 and 4 are supported between levels.  We use
 MPI to distribute AMR grids across nodes and use logical tiling with
 OpenMP to divide a grid across threads for multi-core CPU machines
@@ -91,7 +91,8 @@ OpenMP to divide a grid across threads for multi-core CPU machines
 GPU threads on GPU-based machines (fine-grained parallelism).  All of
 the core physics can run on GPUs and have been shown to scale well.
 For performance portability, we use the same source code
-for both CPUs and GPUs, and implement our parallel loops in an abstraction
+for both CPUs and GPUs; additionally, we implement our parallel loops
+in an abstraction
 layer provided by AMReX. An abstract parallel for loop accepts as arguments
 a range of indices and the body of the loop to execute for a given index,
 and the AMReX backend dispatches the work appropriately (e.g., one cell per
@@ -106,30 +107,32 @@ Nyx was to build a simulation code based on a modern,
 well-supported AMR library (BoxLib which evolved to AMReX), using
 unsplit integration techniques.
 The large developer community contributing to AMReX
-helps in Nyx continually gaining optimizations for new architectures.  
+helps in Nyx continually gaining optimizations for new architectures
[…]
