### 2021-06-14 boxlib-codes/nyx#80 @jmsexton03
I went through and polished up the grammar.
