### 2021-06-21 #3068 @teuben
@dfm   when i looked at the new paper proof, there is no reference for Strang splitting(3rd par)  or the word "integalactic" in the 2nd sentence is misspelled. These were fixed in paper.md, but they didn't show up for me in the paper proof just generated.

True to form, I looked at the PDF generated again, and now it's good.... perhaps I looked too soon.  Sorry for the confusion.
