### 2021-06-21 #3068 @c-white
@teuben I had the same issue. I'm now 90% sure there's a caching issue on whedon's end. While it might instantly generate a new pdf somewhere, it uses the same link as before, and it takes time for the link to point to the updated document.
