### 2021-06-21 AMReX-Astro/Nyx#81 @teuben
paper.md:   references to other codes?
In the "Statement of Need" section it described well the unique aspects of the code, but actually doesn't name a few codes that are similar to Nyx.   For somebody coming is rather fresh and looking at Nyx , it would be beneficial to name a few likewise programs. I liken this to the "SEE ALSO" section in unix man pages, which i find terribly useful to expand my unix knowledge or thinking of other way to achieve the same.
I realize we can't look in the future, this is github, so future codes can even be mentioned here in the future :-)

From openjournals/joss-reviews#3068
