### 2021-06-21 e2f0f04
Add citations for cosmo codes
