+++ Docs/paper/paper.bib
@@ -174,3 +174,209 @@ journal = "Journal of Parallel and Distributed Computing ",
     number = {},
      pages = {71-81}
 }
+
+@article{Bryan2014,
+   abstract = {This paper describes the open-source code Enzo, which uses block-structured adaptive mesh refinement to provide high spatial and temporal resolution for modeling astrophysical fluid flows. The code is Cartesian, can be run in one, two, and three dimensions, and supports a wide variety of physics including hydrodynamics, ideal and non-ideal magnetohydrodynamics, N-body dynamics (and, more broadly, self-gravity of fluids and particles), primordial gas chemistry, optically thin radiative cooling of primordial and metal-enriched plasmas (as well as some optically-thick cooling models), radiation transport, cosmological expansion, and models for star formation and feedback in a cosmological context. In addition to explaining the algorithms implemented, we present solutions for a wide range of test problems, demonstrate the code's parallel performance, and discuss the Enzo collaboration's code development methodology. © 2014. The American Astronomical Society. All rights reserved..},
+   author = {Greg L. Bryan and Michael L. Norman and Brian W. O'Shea and Tom Abel and John H. Wise and Matthew J. Turk and Daniel R. Reynolds and David C. Collins and Peng Wang and Samuel W. Skillman and Britton Smith and Robert P. Harkness and James Bordner and Ji Hoon Kim and Michael Kuhlen and Hao Xu and Nathan Goldbaum and Cameron Hummels and Alexei G. Kritsuk and Elizabeth Tasker and Stephen Skory and Christine M. Simpson and Oliver Hahn and Jeffrey S. Oishi and Geoffrey C. So and Fen Zhao and Renyue Cen and Yuan Li},
+   doi = {10.1088/0067-0049/211/2/19},
+   issn = {00670049},
+   issue = {2},
+   journal = {Astrophysical Journal, Supplement Series},
+   keywords = {hydrodynamics methods,numerical,words},
+   pages = {19},
+   publisher = {Institute of Physics Publishing},
+   title = {Enzo: An adaptive mesh refinement code for astrophysics},
+   volume = {211},
+   url = {http://enzo-project.org},
+   year = {2014},
+}
+@article{Fryxell2000,
+   abstract = {We report on the completion of the first version of a new-generation simulation code, FLASH. The FLASH code solves the fully compressible, reactive hydrodynamic equations and allows for the use of adaptive mesh refinement. It also contains state-of-the-art modules for the equations of state and thermonuclear reaction networks. The FLASH code was developed to study the problems of nuclear flashes on the surfaces of neutron stars and white dwarfs, as well as in the interior of white dwarfs. We expect, however, that the FLASH code will be useful for solving a wide variety of other problems. This first version of the code has been subjected to a large variety of test cases and is currently being used for production simulations of X-ray bursts, Rayleigh-Taylor and Richtmyer-Meshkov instabilities, and thermonuclear flame fronts. The FLASH code is portable and already runs on a wide variety of massively parallel machines, including some of the largest machines now extant.},
+   author = {B. Fryxell and K. Olson and P. Ricker and F. X. Timmes and M. Zingale and D. Q. Lamb and P. MacNeice and R. Rosner and J. W. Truran and H. Tufo},
+   doi = {10.1086/317361},
+   issn = {0067-0049},
+   issue = {1},
+   journal = {The Astrophysical Journal Supplement Series},
+   keywords = {Subject headings : equation of state È hydrodynamics È methods : numerical È nuclear reactions,abundances È stars : general,nucleosynthesis},
+   month = {11},
+   pages = {273-334},
+   publisher = {American Astronomical Society},
+   title = {FLASH: An Adaptive Mesh Hydrodynamics Code for Modeling Astrophysical Thermonuclear Flashes},
+   volume = {131},
[…]
