### 2021-06-21 d410b56
Remove ASCL references and convert non-unicode chars
