### 2021-06-21 e1e31a9
Update bibtex references syntax
