+++ Docs/paper/paper.bib
@@ -176,7 +176,6 @@ journal = "Journal of Parallel and Distributed Computing ",
 }
 
 @article{Bryan2014,
-   abstract = {This paper describes the open-source code Enzo, which uses block-structured adaptive mesh refinement to provide high spatial and temporal resolution for modeling astrophysical fluid flows. The code is Cartesian, can be run in one, two, and three dimensions, and supports a wide variety of physics including hydrodynamics, ideal and non-ideal magnetohydrodynamics, N-body dynamics (and, more broadly, self-gravity of fluids and particles), primordial gas chemistry, optically thin radiative cooling of primordial and metal-enriched plasmas (as well as some optically-thick cooling models), radiation transport, cosmological expansion, and models for star formation and feedback in a cosmological context. In addition to explaining the algorithms implemented, we present solutions for a wide range of test problems, demonstrate the code's parallel performance, and discuss the Enzo collaboration's code development methodology. © 2014. The American Astronomical Society. All rights reserved..},
    author = {Greg L. Bryan and Michael L. Norman and Brian W. O'Shea and Tom Abel and John H. Wise and Matthew J. Turk and Daniel R. Reynolds and David C. Collins and Peng Wang and Samuel W. Skillman and Britton Smith and Robert P. Harkness and James Bordner and Ji Hoon Kim and Michael Kuhlen and Hao Xu and Nathan Goldbaum and Cameron Hummels and Alexei G. Kritsuk and Elizabeth Tasker and Stephen Skory and Christine M. Simpson and Oliver Hahn and Jeffrey S. Oishi and Geoffrey C. So and Fen Zhao and Renyue Cen and Yuan Li},
    doi = {10.1088/0067-0049/211/2/19},
    issn = {00670049},
@@ -229,13 +228,12 @@ journal = "Journal of Parallel and Distributed Computing ",
    journal = {The Astrophysical Journal Supplement Series},
    month = {7},
    pages = {73-94},
-   title = {Adaptive Refinement Tree: A New High‐Resolution <i>N</i> ‐Body Code for Cosmological Simulations},
+   title = {Adaptive Refinement Tree: A New High‐Resolution N‐Body Code for Cosmological Simulations},
    volume = {111},
    url = {https://iopscience.iop.org/article/10.1086/313015},
    year = {1997},
 }
 @inproceedings{Schaller2016,
-   abstract = {We present a new open-source cosmological code, called swift, designed to solve the equations of hydrodynamics using a particle-based approach (Smooth Particle Hydrodynamics) on hybrid shared / distributed-memory architectures. swift was designed from the bottom up to provide excellent strong scaling on both commodity clusters (Tier-2 systems) and Top100-supercomputers (Tier-0 systems), without relying on architecture-specific features or specialized accelerator hardware. This performance is due to three main computational approaches: Task-based parallelism for shared-memory parallelism, which provides fine-grained load balancing and thus strong scaling on large numbers of cores. Graph-based domain decomposition, which uses the task graph to decompose the simulation domain such that the work, as opposed to just the data, as is the case with most partitioning schemes, is equally distributed across all nodes. Fully dynamic and asynchronous communica- Tion, in which communication is modelled as just another task in the task-based scheme, sending data when-ever it is ready and deferring on tasks that rely on data from other nodes until it arrives. In order to use these approaches, the code had to be rewritten from scratch, and the algorithms therein adapted to the task-based paradigm. As a result, we can show upwards of 60% parallel efficiency for moderate-sized problems when increasing the number of cores 512-fold, on both x86-based and Power8-based architectures.},
[…]
