+++ cmake/NyxSetupAMReX.cmake
@@ -99,6 +99,7 @@ else ()
    set(AMReX_SPACEDIM             3                          CACHE INTERNAL "" )
    set(AMReX_PRECISION            DOUBLE                     CACHE INTERNAL "" )
    set(AMReX_FORTRAN              OFF                        CACHE INTERNAL "" )
+   set(AMReX_PROBINIT             OFF                        CACHE INTERNAL "" )
    set(AMReX_MPI                  ${Nyx_MPI}                 CACHE INTERNAL "" )
    set(AMReX_MPI_THREAD_MULTIPLE  ${Nyx_MPI_THREAD_MULTIPLE} CACHE INTERNAL "" )
    set(AMReX_OMP                  ${Nyx_OMP}                 CACHE INTERNAL "" )
