+++ Docs/paper/paper.bib
@@ -332,6 +332,7 @@ Lagrangian perturbation theory.},
    url = {http://arxiv.org/abs/1409.7395 http://dx.doi.org/10.1093/mnras/stv195},
    year = {2014},
 }
+
 @article{Villasenor2021,
    author = {Bruno Villasenor and Brant Robertson and Piero Madau and Evan Schneider},
    doi = {10.3847/1538-4357/abed5a},
@@ -347,6 +348,7 @@ Lagrangian perturbation theory.},
    url = {https://doi.org/10.3847/1538-4357/abed5a},
    year = {2021},
 }
+
 @article{Bordner2018,
    abstract = {Astrophysical and cosmological phenomena involve a large variety of physical processes, and can encompass an enormous range of scales. To effectively investigate these phenomena computationally, applications must similarly support modeling these phenomena on enormous ranges of scales; furthermore, they must do so efficiently on high-performance computing platforms of ever-increasing parallelism and complexity. We describe Enzo-P, a Petascale redesign of the ENZO adaptive mesh refinement astrophysics and cosmology application, along with Cello, a reusable and scalable adaptive mesh refinement software framework, on which Enzo-P is based. Cello's scalability is enabled by the Charm++ Parallel Programming System, whose data-driven asynchronous execution model is ideal for taking advantage of the available but irregular parallelism in adaptive mesh refinement-based applications. We present scaling results on the NSF Blue Waters supercomputer, and outline our future plans to bring Enzo-P to the Exascale Era by targeting highly-heterogeneous accelerator-based platforms.},
    author = {James Bordner and Michael L. Norman},
