### 2021-06-22 3059644
More character encoding fixes
