+++ Docs/paper/paper.bib
@@ -148,7 +148,7 @@ doi = "10.1016/j.jcp.2007.07.035",
 url = "https://doi.org/10.1016/j.jcp.2007.07.035",
 author = "Francesco Miniati and Phillip Colella",
 keywords = "Higher-order Godunov methods, Adaptive mesh refinement, Elliptic methods, Particle-in-cell methods",
-abstract = "We present a new numerical algorithm for the solution of coupled collisional and collisionless systems, based on the block structured adaptive mesh and time refinement strategy (AMR). We describe the issues associated with the discretization of the system equations and the synchronization of the numerical solution on the hierarchy of grid levels. We implement a code based on a higher order, conservative and directionally unsplit Godunov’s method for hydrodynamics; a symmetric, time centered modified symplectic scheme for collisionless component; and a multilevel, multigrid relaxation algorithm for the elliptic equation coupling the two components. Numerical results that illustrate the accuracy of the code and the relative merit of various implemented schemes are also presented."
+abstract = "We present a new numerical algorithm for the solution of coupled collisional and collisionless systems, based on the block structured adaptive mesh and time refinement strategy (AMR). We describe the issues associated with the discretization of the system equations and the synchronization of the numerical solution on the hierarchy of grid levels. We implement a code based on a higher order, conservative and directionally unsplit Godunov's method for hydrodynamics; a symmetric, time centered modified symplectic scheme for collisionless component; and a multilevel, multigrid relaxation algorithm for the elliptic equation coupling the two components. Numerical results that illustrate the accuracy of the code and the relative merit of various implemented schemes are also presented."
 }
 
 @article{Kokkos,
@@ -196,7 +196,6 @@ journal = "Journal of Parallel and Distributed Computing ",
    issn = {0067-0049},
    issue = {1},
    journal = {The Astrophysical Journal Supplement Series},
-   keywords = {Subject headings : equation of state È hydrodynamics È methods : numerical È nuclear reactions,abundances È stars : general,nucleosynthesis},
    month = {11},
    pages = {273-334},
    publisher = {American Astronomical Society},
@@ -206,7 +205,6 @@ journal = "Journal of Parallel and Distributed Computing ",
    year = {2000},
 }
 @article{Teyssier2002,
[…]
