+++ docs_html/_sources/NyxGettingStarted.rst.txt
@@ -29,7 +29,7 @@ the simulations results.
    Compiling Nyx with SUNDIALS 5 <getting_started/NyxSundials>
 
 Additional details about compiling Nyx in different environments can be found
-in the `AMReX build documentation<https://amrex-codes.github.io/amrex/docs_html/BuildingAMReX_Chapter.html>`_.
+in the `AMReX build documentation <https://amrex-codes.github.io/amrex/docs_html/BuildingAMReX_Chapter.html>`_ .
 Containerized Ubuntu builds with CMake are run as 
 `GitHub Actions <https://github.com/AMReX-Astro/Nyx/actions/workflows/linux.yml>`_
 whose workflow files are stored in Nyx/.github/workflows .
