+++ docs_html/_sources/NyxCitation.rst.txt
@@ -6,7 +6,8 @@ If you use Nyx, we appreciate you citing the main code paper:
   * A. S. Almgren, J. B. Bell, M.J. Lijewski, Z. Lukic, E. Van Andel, **"Nyx: A Massively Parallel AMR Code for Computational Cosmology"** Astrophysical Journal, **765**, 39, 2013. [[Journal]](https://iopscience.iop.org/article/10.1088/0004-637X/765/1/39) [[PDF]](https://iopscience.iop.org/article/10.1088/0004-637X/765/1/39/pdf)
 
   In BibTeX::
-  @ARTICLE{Nyx,
+
+    @ARTICLE{Nyx,
                author = {{Almgren}, A.~S. and {Bell}, J.~B. and {Lijewski}, M.~J. and {Luki{\'c}}, Z. and {Van Andel}, E.},
                title = "{Nyx: A Massively Parallel AMR Code for Computational Cosmology}",
                journal = {The Astrophysical Journal},
@@ -19,13 +20,15 @@ If you use Nyx, we appreciate you citing the main code paper:
                eid = {39},
                pages = {39},
                doi = {10.1088/0004-637X/765/1/39}
-  }
+    }
 
 The development of AMReX library is led by the
 Center for Computational Sciences and Engineering / Lawrence Berkeley
 National Laboratory. Nyx development is done collaboratively, including the Computational Cosmology Center and CCSE. 
+
 All of Nyx's development is done in the public github repository—anyone can see the updates as they are done.  
 We are always happy to have new developers as part of the Nyx team. 
+
 Fork the Nyx git repository on github, make your changes, and issue a pull request against the development branch 
 of the github repo. Any level of changes are welcome, including documentation, bug fixes, new test problems, and new solvers.
 
