### 2021-06-22 boxlib-codes/nyx#81 @jmsexton03
Added references to a broad range of similar codes (unigrid / AMR / SPH)
