### 2021-06-22 #3068 @jmsexton03
@dfm Should I make a new github tag whenever changes are final? Did whedon automatically check the doi's like it did for the earlier article proofs? I want to make sure the new references are good too.
