### 2021-06-25 564f990
minor typographic edits for JOSS manuscript
