### 2021-07-02 e0e212c
Update to 21.07
 * Update to 21.07 AMReX submodule
 * Fix sundials namespace include order
 * Make OpenMP+GPU less compiler-specific
