### 2021-07-02 #3068 @jmsexton03
I have created the tagged release: https://github.com/AMReX-Astro/Nyx/releases/tag/21.07
I have published a zenodo release associated with that tag: https://zenodo.org/record/5059767

The incremented tag is 21.07
The zenodo DOI for 21.07 is: 10.5281/zenodo.5059767
The zenodo DOI for "All Versions" is 10.5281/zenodo.5059766
